adminhub policy

+++ app/Http/Controllers/AdminHub/V1/Admin/RoleController.php
@@ -9,11 +9,14 @@ use App\Http\Requests\AdminHub\V1\Admin\RoleUpdateRequest;
 use App\Models\AdminHub\Permission;
 use App\Models\AdminHub\Role;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
     public function index(): JsonResponse
     {
+        Gate::authorize('viewAny', Role::class);
+
         return response()->json([
             'data' => Role::orderBy('sort')->orderBy('id')->get(['id', 'name', 'updated_at'])
                 ->transform(fn ($role) => [
@@ -26,6 +29,8 @@ class RoleController extends Controller
 
     public function create(): JsonResponse
     {
+        Gate::authorize('create', Role::class);
+
         return response()->json([
             'data' => [
                 'actions' => Permission::orderBy('sort')->orderBy('id')->get(['id', 'name', 'action']),
@@ -35,6 +40,8 @@ class RoleController extends Controller
 
     public function store(RoleStoreRequest $request): JsonResponse
     {
+        Gate::authorize('create', Role::class);
+
         $input = $request->safe();
 
         $role = Role::create($input->only('name'));
@@ -54,6 +61,8 @@ class RoleController extends Controller
 
     public function edit(Role $role): JsonResponse
     {
+        Gate::authorize('update', $role);
+
         return response()->json([
             'data' => [
                 'actions' => Permission::orderBy('sort')->orderBy('id')->get(['id', 'name', 'action']),
@@ -63,6 +72,8 @@ class RoleController extends Controller
 
     public function show(Role $role): JsonResponse
     {
+        Gate::authorize('view', $role);
+
         return response()->json([
             'data' => [
                 'id' => $role->id,
@@ -74,6 +85,8 @@ class RoleController extends Controller
 
     public function update(RoleUpdateRequest $request, Role $role): JsonResponse
     {
+        Gate::authorize('update', $role);
+
         $input = $request->safe();
 
         $role->update($input->only('name'));
@@ -95,6 +108,8 @@ class RoleController extends Controller
 
     public function destroy(Role $role): JsonResponse
     {
+        Gate::authorize('delete', $role);
+
         $role->permissions()->detach();
 
         $role->userGroups()->detach();
@@ -106,6 +121,8 @@ class RoleController extends Controller
 
     public function sort(RoleSortRequest $request): JsonResponse
     {
+        Gate::authorize('update', Role::class);
+
         $input = $request->validated();
 
         foreach ($input['ids'] as $key => $id) {
